refactor(angular): move providers to common library (#27899)
