+++ packages/angular/common/src/index.ts
@@ -1,3 +1,21 @@
-// This is required to get ng-packagr to build.
-// Remove this when you actually have something to export
-export const placeholder = true;
+export { ActionSheetController } from './providers/action-sheet-controller';
+export { AlertController } from './providers/alert-controller';
+export { LoadingController } from './providers/loading-controller';
+export { MenuController } from './providers/menu-controller';
+export { ModalController } from './providers/modal-controller';
+export { PickerController } from './providers/picker-controller';
+export { PopoverController } from './providers/popover-controller';
+export { ToastController } from './providers/toast-controller';
+
+export { AnimationController } from './providers/animation-controller';
+export { GestureController } from './providers/gesture-controller';
+export { DomController } from './providers/dom-controller';
+
+export { Config, ConfigToken } from './providers/config';
+export { Platform } from './providers/platform';
+
+export { bindLifecycleEvents, AngularDelegate } from './providers/angular-delegate';
+
+export type { IonicWindow } from './types/interfaces';
+
+export { NavParams } from './directives/navigation/nav-params';
