+++ packages/angular/common/src/providers/animation-controller.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@angular/core';
-import { Animation, createAnimation, getTimeGivenProgression } from '@ionic/core';
+import type { Animation } from '@ionic/core/components';
+import { createAnimation, getTimeGivenProgression } from '@ionic/core/components';
 
 @Injectable({
   providedIn: 'root',
