+++ packages/angular/common/src/providers/picker-controller.ts
@@ -1,7 +1,8 @@
 import { Injectable } from '@angular/core';
-import { PickerOptions, pickerController } from '@ionic/core';
+import type { PickerOptions } from '@ionic/core';
+import { pickerController } from '@ionic/core/components';
 
-import { OverlayBaseController } from '../util/overlay';
+import { OverlayBaseController } from '../utils/overlay';
 
 @Injectable({
   providedIn: 'root',
