+++ packages/angular/common/src/types/interfaces.ts
@@ -7,8 +7,3 @@ export interface IonicWindow extends Window {
   Ionic: IonicGlobal;
   __zone_symbol__requestAnimationFrame?: (ts: FrameRequestCallback) => number;
 }
-
-export interface HTMLStencilElement extends HTMLElement {
-  componentOnReady?(): Promise<this>;
-  forceUpdate?(): void;
-}
