+++ packages/angular/common/src/utils/overlay.ts
