+++ packages/angular/src/directives/navigation/nav-delegate.ts
@@ -1,6 +1,6 @@
 import { ElementRef, Injector, Directive, EnvironmentInjector } from '@angular/core';
+import { AngularDelegate } from '@ionic/angular/common';
 
-import { AngularDelegate } from '../../providers/angular-delegate';
 import { ProxyCmp, proxyOutputs } from '../angular-component-lib/utils';
 
 @ProxyCmp({
