+++ packages/angular/src/providers/nav-controller.ts
@@ -1,11 +1,10 @@
 import { Location } from '@angular/common';
 import { Injectable, Optional } from '@angular/core';
 import { NavigationExtras, Router, UrlSerializer, UrlTree, NavigationStart } from '@angular/router';
+import { Platform } from '@ionic/angular/common';
 import { AnimationBuilder, NavDirection, RouterDirection } from '@ionic/core';
 
 import { IonRouterOutlet } from '../directives/navigation/ion-router-outlet';
-
-import { Platform } from './platform';
 
 export interface AnimationOptions {
   animated?: boolean;
